my_work/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
autogen-agentchat==0.4.9.2
autogen-core==0.4.9.2
autogen-ext==0.4.9.2
autogenstudio==0.4.2.1
azure-ai-documentintelligence==1.0.2
azure-ai-inference==1.0.0b9
azure-core==1.33.0
azure-identity==1.21.0
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.4
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
greenlet==3.2.0
h11==0.14.0
html2text==2025.4.15
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonref==1.1.0
kiwisolver==1.4.8
Levenshtein==0.27.1
lightning-utilities==0.14.3
llvmlite==0.44.0
loguru==0.7.3
lxml==5.3.2
Mako==1.3.10
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.0.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.4.3
networkx==3.4.2
numba==0.61.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
openai==1.74.0
openpyxl==3.1.5
opentelemetry-api==1.32.1
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.3
pdfminer.six==20250416
Pillow==9.3.0
playwright==1.51.0
propcache==0.3.1
protobuf==5.29.4
psycopg==3.2.6
puremagic==1.28
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
python-pptx==1.0.2
pytorch-lightning==2.5.1
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
sympy==1.13.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchmetrics==1.7.1
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xgboost==2.1.3
xlrd==2.0.1
XlsxWriter==3.2.2
yarl==1.20.0
youtube-transcript-api==1.0.3
zipp==3.21.0